# fifocache: shard values map