--- a/pkg/fileservice/fifocache/bench_test.go
+++ b/pkg/fileservice/fifocache/bench_test.go
@@ -20,7 +20,7 @@ import (
 
 func BenchmarkSequentialSet(b *testing.B) {
 	size := 65536
-	cache := New[int, int](size, nil)
+	cache := New[int, int](size, nil, ShardInt[int])
 	nElements := size * 16
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
@@ -30,7 +30,7 @@ func BenchmarkSequentialSet(b *testing.B) {
 
 func BenchmarkParallelSet(b *testing.B) {
 	size := 65536
-	cache := New[int, int](size, nil)
+	cache := New[int, int](size, nil, ShardInt[int])
 	nElements := size * 16
 	b.ResetTimer()
 	b.RunParallel(func(pb *testing.PB) {
@@ -42,7 +42,7 @@ func BenchmarkParallelSet(b *testing.B) {
 
 func BenchmarkGet(b *testing.B) {
 	size := 65536
-	cache := New[int, int](size, nil)
+	cache := New[int, int](size, nil, ShardInt[int])
 	nElements := size * 16
 	for i := 0; i < b.N; i++ {
 		cache.Set(i%nElements, i, 1+i%3)
@@ -55,7 +55,7 @@ func BenchmarkGet(b *testing.B) {
 
 func BenchmarkParallelGet(b *testing.B) {
 	size := 65536
-	cache := New[int, int](size, nil)
+	cache := New[int, int](size, nil, ShardInt[int])
 	nElements := size * 16
 	for i := 0; i < b.N; i++ {
 		cache.Set(i%nElements, i, 1+i%3)
@@ -70,7 +70,7 @@ func BenchmarkParallelGet(b *testing.B) {
 
 func BenchmarkParallelGetOrSet(b *testing.B) {
 	size := 65536
-	cache := New[int, int](size, nil)
+	cache := New[int, int](size, nil, ShardInt[int])
 	nElements := size * 16
 	b.ResetTimer()
 	b.RunParallel(func(pb *testing.PB) {