## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #12506

## What this PR does / why we need it:
performance improved

```
goos: linux
goarch: amd64
pkg: github.com/matrixorigin/matrixone/pkg/fileservice/fifocache
cpu: AMD Ryzen 9 7900 12-Core Processor             
                    │      b1       │                  b2                   │
                    │    sec/op     │    sec/op     vs base                 │
SequentialSet-24       123.7n ± ∞ ¹   133.4n ± ∞ ¹        ~ (p=0.100 n=3) ²
ParallelSet-24        106.80n ± ∞ ¹   26.50n ± ∞ ¹        ~ (p=0.100 n=3) ²
Get-24                 17.54n ± ∞ ¹   32.77n ± ∞ ¹        ~ (p=0.100 n=3) ²
ParallelGet-24        41.450n ± ∞ ¹   5.739n ± ∞ ¹        ~ (p=0.100 n=3) ²
ParallelGetOrSet-24    28.53n ± ∞ ¹   29.40n ± ∞ ¹        ~ (p=0.200 n=3) ²
geomean                48.70n         28.72n        -41.03%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                    │     b1      │                 b2                  │
                    │    B/op     │    B/op      vs base                │
SequentialSet-24      32.00 ± ∞ ¹   32.00 ± ∞ ¹       ~ (p=1.000 n=3) ²
ParallelSet-24        1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
Get-24                0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
ParallelGet-24        0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
ParallelGetOrSet-24   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
geomean                         ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                    │     b1      │                 b2                  │
                    │  allocs/op  │  allocs/op   vs base                │
SequentialSet-24      1.000 ± ∞ ¹   1.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
ParallelSet-24        0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
Get-24                0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
ParallelGet-24        0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
ParallelGetOrSet-24   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=3) ²
geomean                         ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
```